requirements.txt
# Web app
streamlit>=1.34.0

# Supabase — auth + pgvector storage
supabase>=2.3.0

# Groq LLM API
groq>=1.0.0

# LangChain — only the sub-packages actually used
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0

# Embeddings
sentence-transformers>=3.0.0

# PDF parsing
PyMuPDF>=1.25.0

# DOCX parsing
python-docx>=1.1.0

# Environment variables
python-dotenv>=1.0.0

# PDF resume generation
reportlab>=4.0.0

# Charts
plotly>=5.0.0

# HTTP (GitHub OAuth + JSearch API)
requests>=2.31.0

# KeyBERT — BERT-based keyword extraction
keybert>=0.8.0

# PCA embedding visualization
scikit-learn>=1.3.0

# BM25 sparse retrieval (hybrid RAG)
rank-bm25>=0.2.2
